+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 * `Details`/`Reason`/`NewDetails` field names in reviewable request types unified to `creatorDetails`
 * Corresponding messages in `messages` map
-
+* Calculated fee response
 ### Removed
 
 * `UpdateSaleEndTimeRequest` type